requirements.txt
osmnx
matplotlib
pillow
numpy
